refactor(components/molecule/tabs): index files and refactor
