+++ components/molecule/tabs/src/config.js
@@ -1,0 +1,15 @@
+export const BASE_CLASS = `sui-MoleculeTabs`
+
+export const CLASS_SCROLLER = `${BASE_CLASS}-scroller`
+export const CLASS_CONTENT = `${BASE_CLASS}-content`
+
+export const TYPES = {
+  HORIZONTAL: 'horizontal',
+  VERTICAL: 'vertical',
+  FULLWIDTH: 'fullWidth'
+}
+
+export const VARIANTS = {
+  HIGHLIGHTED: 'highlighted',
+  CLASSIC: 'classic'
+}
